+++ lib/td_cache/application.ex
@@ -11,12 +11,17 @@ defmodule TdCache.Application do
 
     children =
       [
+        {Redix, host: redis_host, port: port, name: :redix},
         {TdCache.Redix.Pool, redis_host: redis_host, port: port},
         TdCache.ConceptCache,
         TdCache.TemplateCache,
+        TdCache.UserCache,
+        TdCache.TaxonomyCache,
         con_cache_child_spec(:templates, 10, 60),
+        con_cache_child_spec(:users, 10, 60),
+        con_cache_child_spec(:taxonomy, 10, 60),
         con_cache_child_spec(:concepts, 10, 60)
-      ] ++ cache_cleaner_workers() ++ event_stream_workers()
+      ] ++ cache_cleaner_workers() ++ event_stream_workers(redis_host, port)
 
     # See https://hexdocs.pm/elixir/Supervisor.html
     # for other strategies and supported options
@@ -24,10 +29,13 @@ defmodule TdCache.Application do
     Supervisor.start_link(children, opts)
   end
 
-  defp event_stream_workers do
+  defp event_stream_workers(redis_host, port) do
     case Application.get_env(:td_cache, :event_stream) do
-      nil -> []
-      config -> [{TdCache.EventStream, config}]
+      nil ->
+        []
+
+      config ->
+        [{TdCache.EventStream, Keyword.merge([redis_host: redis_host, port: port], config)}]
     end
   end